+++ metrics/src/recorder.rs
@@ -48,7 +48,7 @@ pub trait Recorder {
     fn update_gauge(&self, key: Key, value: GaugeValue);
 
     /// Records a histogram.
-    fn record_histogram(&self, key: Key, value: u64);
+    fn record_histogram(&self, key: Key, value: f64);
 }
 
 /// A no-op recorder.
@@ -70,7 +70,7 @@ impl Recorder for NoopRecorder {
     }
     fn increment_counter(&self, _key: Key, _value: u64) {}
     fn update_gauge(&self, _key: Key, _value: GaugeValue) {}
-    fn record_histogram(&self, _key: Key, _value: u64) {}
+    fn record_histogram(&self, _key: Key, _value: f64) {}
 }
 
 /// Sets the global recorder to a `&'static Recorder`.
